+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.API.Tests/Integration/Tasks/KickoffMeetingTaskApiTests.cs
@@ -1,7 +1,9 @@
 ﻿using Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Tasks;
 using Dfe.ManageFreeSchoolProjects.API.Tests.Fixtures;
 using Dfe.ManageFreeSchoolProjects.API.Tests.Helpers;
+using Dfe.ManageFreeSchoolProjects.API.Constants;
 using System;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace Dfe.ManageFreeSchoolProjects.API.Tests.Integration.Tasks
@@ -42,6 +44,10 @@ namespace Dfe.ManageFreeSchoolProjects.API.Tests.Integration.Tasks
             projectResponse.KickOffMeeting.RealisticYearOfOpening.Should().Be(request.KickOffMeeting.RealisticYearOfOpening);
             projectResponse.KickOffMeeting.FundingArrangementAgreed.Should().Be(request.KickOffMeeting.FundingArrangementAgreed);
             projectResponse.KickOffMeeting.FundingArrangementDetailsAgreed.Should().Be(request.KickOffMeeting.FundingArrangementDetailsAgreed);
+
+            using var contextPostSave = _testFixture.GetContext();
+            var updatedProject = contextPostSave.Kpi.First(p => p.ProjectStatusProjectId == projectId);
+            updatedProject.RyooWd.Should().Be(request.KickOffMeeting.RealisticYearOfOpening);
         }
 
         [Fact]
@@ -77,6 +83,49 @@ namespace Dfe.ManageFreeSchoolProjects.API.Tests.Integration.Tasks
             projectResponse.KickOffMeeting.RealisticYearOfOpening.Should().Be(request.KickOffMeeting.RealisticYearOfOpening);
             projectResponse.KickOffMeeting.FundingArrangementAgreed.Should().Be(request.KickOffMeeting.FundingArrangementAgreed);
             projectResponse.KickOffMeeting.FundingArrangementDetailsAgreed.Should().Be(request.KickOffMeeting.FundingArrangementDetailsAgreed);
+
+            using var contextPostSave = _testFixture.GetContext();
+            var updatedProject = contextPostSave.Kpi.First(p => p.ProjectStatusProjectId == projectId);
+            updatedProject.RyooWd.Should().Be(request.KickOffMeeting.RealisticYearOfOpening);
+        }
+
+        [Fact]
+        public async Task Patch_ExistingKickOffMeeting_NoRYOOSetsDefaultValue()
+        {
+            var project = DatabaseModelBuilder.BuildProject();
+            var projectId = project.ProjectStatusProjectId;
+
+            using var context = _testFixture.GetContext();
+            context.Kpi.Add(project);
+
+            var kickOffMeetingTask = DatabaseModelBuilder.BuildKickOffMeetingTask(project.Rid);
+            context.Milestones.Add(kickOffMeetingTask);
+
+            await context.SaveChangesAsync();
+
+            var dateNineDaysInFuture = new DateTime().AddDays(9);
+
+            var request = new UpdateProjectByTaskRequest()
+            {
+                KickOffMeeting = new KickOffMeetingTask()
+                {
+                    FundingArrangementAgreed = true,
+                    RealisticYearOfOpening = "",
+                    FundingArrangementDetailsAgreed = "text",
+                    SavedDocumentsInWorkplacesFolder = true
+                }
+            };
+
+            var projectResponse = await _client.UpdateProjectTask(projectId, request, TaskName.KickOffMeeting.ToString());
+
+            projectResponse.KickOffMeeting.SavedDocumentsInWorkplacesFolder.Should().Be(request.KickOffMeeting.SavedDocumentsInWorkplacesFolder);
+            projectResponse.KickOffMeeting.RealisticYearOfOpening.Should().Be(request.KickOffMeeting.RealisticYearOfOpening);
+            projectResponse.KickOffMeeting.FundingArrangementAgreed.Should().Be(request.KickOffMeeting.FundingArrangementAgreed);
+            projectResponse.KickOffMeeting.FundingArrangementDetailsAgreed.Should().Be(request.KickOffMeeting.FundingArrangementDetailsAgreed);
+
+            using var contextPostSave = _testFixture.GetContext();
+            var updatedProject = contextPostSave.Kpi.First(p => p.ProjectStatusProjectId == projectId);
+            updatedProject.RyooWd.Should().Be(ProjectConstants.RYOODefaultValue);
         }
     }
 }